+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 druid CHANGELOG
 ===============
 
+0.0.3
+-----
+* Updated default Druid version to 0.7.0
+* Moved archive filename spec to attributes
+* Now use new common config file where appropriate
+* Now checkout druid source code from github and build using maven
+
 0.0.2
 -----
 * Replaced symbolic links with regular files